refactor(orc8r): Segregation of Hello Servicer as External
